Retrieve the gif through the Giphy API
The gif is retrieved through the Giphy API, using a public API key.
see https://developers.giphy.com/docs/api#quick-start-guide

The fetch function is used with the Giphy API URL. This returns
a promise which resolves in an object with the representation of
the entire HTTP response to the fetch request.

The .json() method is used to extract the actual data. This returns
a nother promise which resolves in an object ('response') with the
representation of the entire required data.

We need the response.data.images.original.url data to be used as
the src of the img element to display the image.