Add spin animation while the gif is being refreshed
Since a FontAwesome icon is used for the refresh button symbol,
the spin animation can be simply applied by toggling the '.fa-spin'
class.

Note that we want the refresh button icon to spin until the new gif
has been loaded (and not just untild the gif data have been
retrieved).

Then, instead of changing the 'src' field of the <img> element
immediately, a new function setImgSrc(src) has been created which
returns a Promise, which in turn sets the 'src' attribute and
resolves once the image has been loaded.

Then, the '.fa-spin' class is removed from the refresh button icon.
This is done in the '.finally()' method, instead of '.then()', to
avoid the symbol to spin indefinitely in case of errors.