Add credit footer
A local module creditFooter.js has been imported to setup the
credit footer at the bottom of the page. This is a template which
creates such credit footer and style it programmatically.

This module requires importing another local module
fontAwesomeUtilities.js (a wrapper for handling fontAwesome icons
via js).

The WebFont library is also required to load the fonts used in
this footer via js.
see https://github.com/typekit/webfontloader

Since this project does not use npm, these scripts are loaded in
the <head> of index.html.